add binary response example
